+++ Source/Stazor.Core/IStazorLogger.cs
@@ -8,9 +8,4 @@
     public interface IStazorLogger<out TCategoryName> : IStazorLogger
     {
     }
-
-    public interface IStazorLoggerFactory
-    {
-        IStazorLogger CreateLogger<TCategory>();
-    }
 }
