+++ Source/Stazor.Core/IStazorLoggerFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace Stazor.Core
+{
+    public interface IStazorLoggerFactory
+    {
+        IStazorLogger CreateLogger<TCategoryName>();
+    }
+}
